# Vectorize `search` for 32-bit and 64-bit elements, also improve 8-bit and 16-bit vectorization