--- a/benchmarks/src/search.cpp
+++ b/benchmarks/src/search.cpp
@@ -52,8 +52,9 @@ constexpr data_and_pattern patterns[] = {
 template <class T>
 using not_highly_aligned_basic_string = std::basic_string<T, std::char_traits<T>, not_highly_aligned_allocator<T>>;
 
-using not_highly_aligned_string  = not_highly_aligned_basic_string<char>;
-using not_highly_aligned_wstring = not_highly_aligned_basic_string<wchar_t>;
+using not_highly_aligned_string    = not_highly_aligned_basic_string<char>;
+using not_highly_aligned_wstring   = not_highly_aligned_basic_string<wchar_t>;
+using not_highly_aligned_u32string = not_highly_aligned_basic_string<char32_t>;
 
 void c_strstr(benchmark::State& state) {
     const auto& src_haystack = patterns[static_cast<size_t>(state.range())].data;
@@ -205,6 +206,7 @@ BENCHMARK(search_default_searcher<std::uint64_t>)->Apply(common_args);
 
 BENCHMARK(member_find<not_highly_aligned_string>)->Apply(common_args);
 BENCHMARK(member_find<not_highly_aligned_wstring>)->Apply(common_args);
+BENCHMARK(member_find<not_highly_aligned_u32string>)->Apply(common_args);
 
 BENCHMARK(classic_find_end<std::uint8_t>)->Apply(common_args);
 BENCHMARK(classic_find_end<std::uint16_t>)->Apply(common_args);
@@ -218,5 +220,6 @@ BENCHMARK(ranges_find_end<std::uint64_t>)->Apply(common_args);
 
 BENCHMARK(member_rfind<not_highly_aligned_string>)->Apply(common_args);
 BENCHMARK(member_rfind<not_highly_aligned_wstring>)->Apply(common_args);
+BENCHMARK(member_rfind<not_highly_aligned_u32string>)->Apply(common_args);
 
 BENCHMARK_MAIN();