@fivemeyestore 2025-05-10
Your approach to this problem is very elegant. I'm learning a lot from reviewing your code.