@fivemhostingserver 2025-05-10
The performance optimizations here are quite clever. Nice engineering work!