@myeyestore 2025-05-10
Impressive project structure! The architecture really demonstrates solid design principles.