@AlexGuteniev 2025-05-11
I'd be curious about benchmark results for this new `search`/`find_end` vectorization compared to `main` on other CPUs. AMDs is of most interest, but Intels of significantly higher or significantly  lower grade are also of certain interest. Mine is `12th Gen Intel(R) Core(TM) i5-1235U`.

Only one of algorithm type is of interest, so but for all element sizes, also `strstr` is still an insteresting baseline, so running `out\bench\benchmark-search.exe --benchmark_filter="strstr|classic_"` is a good way to filter in only what is needed.

I've modified the benchmarks before the actual changes, so commit 1fb4bfe is a convenient point for benchmarking the "before" state.

<details>
<summary>My own results for E-core affinity</summary>

Benchmark                         |    Before   |   After      | Speedup
----------------------------------|-------------|--------------|--------
c_strstr/0                        |    775 ns   |      772 ns  |  1.00
c_strstr/1                        |    854 ns   |      854 ns  |  1.00
c_strstr/2                        |   69.6 ns   |     69.7 ns  |  1.00
c_strstr/3                        |   37.9 ns   |     38.0 ns  |  1.00
c_strstr/4                        |   3091 ns   |     3098 ns  |  1.00
c_strstr/5                        |  48235 ns   |    48337 ns  |  1.00
classic_search<std::uint8_t>/0    |    771 ns   |      545 ns  |  1.41
classic_search<std::uint8_t>/1    |    832 ns   |      675 ns  |  1.23
classic_search<std::uint8_t>/2    |   78.9 ns   |     27.2 ns  |  2.90
classic_search<std::uint8_t>/3    |   34.2 ns   |     22.0 ns  |  1.55
classic_search<std::uint8_t>/4    |   2151 ns   |     7439 ns  |  0.29
classic_search<std::uint8_t>/5    |  20011 ns   |     7297 ns  |  2.74
classic_search<std::uint16_t>/0   |   1656 ns   |      494 ns  |  3.35
classic_search<std::uint16_t>/1   |   1679 ns   |      429 ns  |  3.91
classic_search<std::uint16_t>/2   |    148 ns   |     34.5 ns  |  4.29
classic_search<std::uint16_t>/3   |   61.1 ns   |     20.5 ns  |  2.98
classic_search<std::uint16_t>/4   |  10407 ns   |     6029 ns  |  1.73
classic_search<std::uint16_t>/5   |  26230 ns   |    18793 ns  |  1.40
classic_search<std::uint32_t>/0   |   3488 ns   |      529 ns  |  6.59
classic_search<std::uint32_t>/1   |   3843 ns   |      614 ns  |  6.26
classic_search<std::uint32_t>/2   |    297 ns   |     36.5 ns  |  8.14
classic_search<std::uint32_t>/3   |    101 ns   |     27.2 ns  |  3.71
classic_search<std::uint32_t>/4   |  13720 ns   |     5782 ns  |  2.37
classic_search<std::uint32_t>/5   |  28629 ns   |    22707 ns  |  1.26
classic_search<std::uint64_t>/0   |   3872 ns   |     1158 ns  |  3.34
classic_search<std::uint64_t>/1   |   4229 ns   |     1271 ns  |  3.33
classic_search<std::uint64_t>/2   |    301 ns   |     94.6 ns  |  3.18
classic_search<std::uint64_t>/3   |    109 ns   |     44.9 ns  |  2.43
classic_search<std::uint64_t>/4   |  19182 ns   |    14350 ns  |  1.34
classic_search<std::uint64_t>/5   |  42780 ns   |    23834 ns  |  1.79
classic_find_end<std::uint8_t>/0  |   52.4 ns   |     33.6 ns  |  1.56
classic_find_end<std::uint8_t>/1  |   47.6 ns   |     32.2 ns  |  1.48
classic_find_end<std::uint8_t>/2  |    747 ns   |      126 ns  |  5.93
classic_find_end<std::uint8_t>/3  |    871 ns   |      157 ns  |  5.55
classic_find_end<std::uint8_t>/4  |   7467 ns   |     5302 ns  |  1.41
classic_find_end<std::uint8_t>/5  |  29309 ns   |     5186 ns  |  5.65
classic_find_end<std::uint16_t>/0 |   94.7 ns   |     36.1 ns  |  2.62
classic_find_end<std::uint16_t>/1 |   89.2 ns   |     28.2 ns  |  3.16
classic_find_end<std::uint16_t>/2 |   1543 ns   |      176 ns  |  8.77
classic_find_end<std::uint16_t>/3 |   1816 ns   |      292 ns  |  6.22
classic_find_end<std::uint16_t>/4 |  14271 ns   |     4653 ns  |  3.07
classic_find_end<std::uint16_t>/5 |  31692 ns   |    17964 ns  |  1.76
classic_find_end<std::uint32_t>/0 |    139 ns   |     30.5 ns  |  4.56
classic_find_end<std::uint32_t>/1 |    136 ns   |     50.0 ns  |  2.72
classic_find_end<std::uint32_t>/2 |   2433 ns   |      365 ns  |  6.67
classic_find_end<std::uint32_t>/3 |   2890 ns   |      509 ns  |  5.68
classic_find_end<std::uint32_t>/4 |  12794 ns   |     4451 ns  |  2.87
classic_find_end<std::uint32_t>/5 |  28414 ns   |    19038 ns  |  1.49
classic_find_end<std::uint64_t>/0 |    177 ns   |     55.2 ns  |  3.21
classic_find_end<std::uint64_t>/1 |    174 ns   |     55.9 ns  |  3.11
classic_find_end<std::uint64_t>/2 |   2905 ns   |      915 ns  |  3.17
classic_find_end<std::uint64_t>/3 |   3441 ns   |     1099 ns  |  3.13
classic_find_end<std::uint64_t>/4 |  19198 ns   |    13903 ns  |  1.38
classic_find_end<std::uint64_t>/5 |  42716 ns   |    23509 ns  |  1.82
</details>