@StephanTLavavej 2025-05-16
<details><summary>Click to expand 5950X results:</summary>

Benchmark                                      |   Before |    After | Speedup
-----------------------------------------------|---------:|---------:|--------
`c_strstr/0`                                   |   141 ns |   138 ns | 1.02
`c_strstr/1`                                   |   159 ns |   155 ns | 1.03
`c_strstr/2`                                   |  15.3 ns |  14.7 ns | 1.04
`c_strstr/3`                                   |  9.60 ns |  9.27 ns | 1.04
`c_strstr/4`                                   |  1062 ns |  1058 ns | 1.00
`c_strstr/5`                                   | 12351 ns | 11788 ns | 1.05
`classic_search<std::uint8_t>/0`               |   158 ns |   202 ns | 0.78
`classic_search<std::uint8_t>/1`               |   173 ns |   229 ns | 0.76
`classic_search<std::uint8_t>/2`               |  20.6 ns |  16.6 ns | 1.24
`classic_search<std::uint8_t>/3`               |  12.1 ns |  13.3 ns | 0.91
`classic_search<std::uint8_t>/4`               |  1016 ns |  3058 ns | 0.33
`classic_search<std::uint8_t>/5`               |  8219 ns |  2966 ns | 2.77
`classic_search<std::uint16_t>/0`              |   307 ns |   223 ns | 1.38
`classic_search<std::uint16_t>/1`              |   348 ns |   210 ns | 1.66
`classic_search<std::uint16_t>/2`              |  33.4 ns |  19.6 ns | 1.70
`classic_search<std::uint16_t>/3`              |  19.6 ns |  12.7 ns | 1.54
`classic_search<std::uint16_t>/4`              |  5003 ns |  2530 ns | 1.98
`classic_search<std::uint16_t>/5`              | 13542 ns |  9725 ns | 1.39
`classic_search<std::uint32_t>/0`              |  1967 ns |   283 ns | 6.95
`classic_search<std::uint32_t>/1`              |  2144 ns |   344 ns | 6.23
`classic_search<std::uint32_t>/2`              |   172 ns |  22.6 ns | 7.61
`classic_search<std::uint32_t>/3`              |  70.2 ns |  16.6 ns | 4.23
`classic_search<std::uint32_t>/4`              |  6325 ns |  2320 ns | 2.73
`classic_search<std::uint32_t>/5`              | 16600 ns |  9653 ns | 1.72
`classic_search<std::uint64_t>/0`              |  1372 ns |   472 ns | 2.91
`classic_search<std::uint64_t>/1`              |  1500 ns |   521 ns | 2.88
`classic_search<std::uint64_t>/2`              |   119 ns |  39.3 ns | 3.03
`classic_search<std::uint64_t>/3`              |  51.2 ns |  21.6 ns | 2.37
`classic_search<std::uint64_t>/4`              |  6645 ns |  9358 ns | 0.71
`classic_search<std::uint64_t>/5`              | 16154 ns | 13946 ns | 1.16
`member_find<not_highly_aligned_string>/0`     |   156 ns |   203 ns | 0.77
`member_find<not_highly_aligned_string>/1`     |   173 ns |   228 ns | 0.76
`member_find<not_highly_aligned_string>/2`     |  20.7 ns |  17.0 ns | 1.22
`member_find<not_highly_aligned_string>/3`     |  12.5 ns |  13.0 ns | 0.96
`member_find<not_highly_aligned_string>/4`     |  1013 ns |  3056 ns | 0.33
`member_find<not_highly_aligned_string>/5`     |  8181 ns |  2965 ns | 2.76
`member_find<not_highly_aligned_wstring>/0`    |   307 ns |   223 ns | 1.38
`member_find<not_highly_aligned_wstring>/1`    |   347 ns |   211 ns | 1.64
`member_find<not_highly_aligned_wstring>/2`    |  34.1 ns |  20.4 ns | 1.67
`member_find<not_highly_aligned_wstring>/3`    |  20.1 ns |  13.6 ns | 1.48
`member_find<not_highly_aligned_wstring>/4`    |  4981 ns |  2514 ns | 1.98
`member_find<not_highly_aligned_wstring>/5`    | 13623 ns |  9676 ns | 1.41
`member_find<not_highly_aligned_u32string>/0`  |  1143 ns |   283 ns | 4.04
`member_find<not_highly_aligned_u32string>/1`  |  1226 ns |   346 ns | 3.54
`member_find<not_highly_aligned_u32string>/2`  |  92.6 ns |  23.0 ns | 4.03
`member_find<not_highly_aligned_u32string>/3`  |  37.3 ns |  17.2 ns | 2.17
`member_find<not_highly_aligned_u32string>/4`  |  7569 ns |  2315 ns | 3.27
`member_find<not_highly_aligned_u32string>/5`  | 17833 ns |  9722 ns | 1.83
`classic_find_end<std::uint8_t>/0`             |  17.8 ns |  19.7 ns | 0.90
`classic_find_end<std::uint8_t>/1`             |  17.1 ns |  19.6 ns | 0.87
`classic_find_end<std::uint8_t>/2`             |   192 ns |  71.9 ns | 2.67
`classic_find_end<std::uint8_t>/3`             |   236 ns |  86.3 ns | 2.73
`classic_find_end<std::uint8_t>/4`             |  6726 ns |  3131 ns | 2.15
`classic_find_end<std::uint8_t>/5`             | 19026 ns |  2869 ns | 6.63
`classic_find_end<std::uint16_t>/0`            |  28.2 ns |  22.0 ns | 1.28
`classic_find_end<std::uint16_t>/1`            |  27.9 ns |  15.7 ns | 1.78
`classic_find_end<std::uint16_t>/2`            |   435 ns |   107 ns | 4.07
`classic_find_end<std::uint16_t>/3`            |   484 ns |   154 ns | 3.14
`classic_find_end<std::uint16_t>/4`            | 13612 ns |  2248 ns | 6.06
`classic_find_end<std::uint16_t>/5`            | 24622 ns |  9311 ns | 2.64
`classic_find_end<std::uint32_t>/0`            |   108 ns |  14.9 ns | 7.25
`classic_find_end<std::uint32_t>/1`            |   102 ns |  20.9 ns | 4.88
`classic_find_end<std::uint32_t>/2`            |  1834 ns |   172 ns | 10.66
`classic_find_end<std::uint32_t>/3`            |  2139 ns |   294 ns | 7.28
`classic_find_end<std::uint32_t>/4`            |  6390 ns |  2048 ns | 3.12
`classic_find_end<std::uint32_t>/5`            | 16613 ns |  9647 ns | 1.72
`classic_find_end<std::uint64_t>/0`            |   107 ns |  34.5 ns | 3.10
`classic_find_end<std::uint64_t>/1`            |   102 ns |  37.0 ns | 2.76
`classic_find_end<std::uint64_t>/2`            |  1845 ns |   498 ns | 3.70
`classic_find_end<std::uint64_t>/3`            |  2144 ns |   552 ns | 3.88
`classic_find_end<std::uint64_t>/4`            |  5712 ns |  8877 ns | 0.64
`classic_find_end<std::uint64_t>/5`            | 15608 ns | 13323 ns | 1.17
`member_rfind<not_highly_aligned_string>/0`    |  18.4 ns |  19.9 ns | 0.92
`member_rfind<not_highly_aligned_string>/1`    |  17.8 ns |  20.0 ns | 0.89
`member_rfind<not_highly_aligned_string>/2`    |   193 ns |  72.4 ns | 2.67
`member_rfind<not_highly_aligned_string>/3`    |   237 ns |  86.7 ns | 2.73
`member_rfind<not_highly_aligned_string>/4`    |  6729 ns |  2742 ns | 2.45
`member_rfind<not_highly_aligned_string>/5`    | 19110 ns |  2640 ns | 7.24
`member_rfind<not_highly_aligned_wstring>/0`   |  28.7 ns |  21.8 ns | 1.32
`member_rfind<not_highly_aligned_wstring>/1`   |  28.5 ns |  16.3 ns | 1.75
`member_rfind<not_highly_aligned_wstring>/2`   |   435 ns |   107 ns | 4.07
`member_rfind<not_highly_aligned_wstring>/3`   |   484 ns |   157 ns | 3.08
`member_rfind<not_highly_aligned_wstring>/4`   | 13626 ns |  2247 ns | 6.06
`member_rfind<not_highly_aligned_wstring>/5`   | 24005 ns |  9375 ns | 2.56
`member_rfind<not_highly_aligned_u32string>/0` |  79.5 ns |  15.9 ns | 5.00
`member_rfind<not_highly_aligned_u32string>/1` |  76.2 ns |  21.6 ns | 3.53
`member_rfind<not_highly_aligned_u32string>/2` |  1283 ns |   173 ns | 7.42
`member_rfind<not_highly_aligned_u32string>/3` |  1507 ns |   296 ns | 5.09
`member_rfind<not_highly_aligned_u32string>/4` |  8882 ns |  2052 ns | 4.33
`member_rfind<not_highly_aligned_u32string>/5` | 27068 ns |  9658 ns | 2.80
</details>

Ignoring the /4 and /5 "evil" cases, I'm somewhat concerned about perf regressions for 1-byte search.